Use records for commands
