+++ src/TrackProgress/SFA.DAS.TrackProgress.Api.AcceptanceTests/ModelTests/UkprnTests.cs
@@ -7,7 +7,7 @@ public class UkprnTests
 {
 	[Test]
 	public void Parse_a_valid_ukprn()
-		=> UkPrn.Parse("123456").Should().Be(123456);
+		=> Ukprn.Parse("123456")!.Value.Should().Be(123456);
 
     [TestCase(null)]
     [TestCase("")]
@@ -15,5 +15,5 @@ public class UkprnTests
     [TestCase("9223372036854775808")]
     [TestCase("-1")]
     public void Parse_an_invalid_ukprn(string input)
-        => UkPrn.Parse(input).Should().Be(0);
+        => Ukprn.Parse(input).Should().BeNull();
 }
