+++ src/TrackProgress/SFA.DAS.TrackProgress.Api/AppStart/AddServiceRegistrationExtension.cs
@@ -2,6 +2,7 @@
 using SFA.DAS.Api.Common.Interfaces;
 using SFA.DAS.SharedOuterApi.Infrastructure;
 using SFA.DAS.SharedOuterApi.Interfaces;
+using SFA.DAS.TrackProgress.Application.Services;
 
 namespace SFA.DAS.TrackProgress.Api.AppStart;
