+++ src/TrackProgress/SFA.DAS.TrackProgress/Apis/CommitmentsV2InnerApi/GetProviderResponse.cs
@@ -1,12 +1,10 @@
-﻿using System;
-using System.Net;
+﻿using System.Net;
 
-namespace SFA.DAS.TrackProgress.Apis.CommitmentsV2InnerApi
+namespace SFA.DAS.TrackProgress.Apis.CommitmentsV2InnerApi;
+
+public class GetProviderResponse
 {
-    public class GetProviderResponse
-    {
-        public HttpStatusCode StatusCode { get; set; }
-        public long ProviderId { get; set; }
-        public string Name { get; set; } = string.Empty;
-    }
+    public HttpStatusCode StatusCode { get; set; }
+    public long ProviderId { get; set; }
+    public string Name { get; set; } = string.Empty;
 }
