+++ src/TrackProgress/SFA.DAS.TrackProgress/Application/DTOs/ProgressDto.cs
@@ -1,8 +1,7 @@
-﻿namespace SFA.DAS.TrackProgress.Application.DTOs
+﻿namespace SFA.DAS.TrackProgress.Application.DTOs;
+
+public class ProgressDto
 {
-    public class ProgressDto
-    {
-        public long ApprovalId { get; set; }
-        public bool OnTrack { get; set; }
-    }
+    public long ApprovalId { get; set; }
+    public bool OnTrack { get; set; }
 }
