+++ src/TrackProgress/SFA.DAS.TrackProgress/Application/Models/Ukprn.cs
@@ -1,11 +1,15 @@
 ﻿namespace SFA.DAS.TrackProgress.Application.Models;
 
-public record UkPrn
+public record Ukprn
 {
-	public static long Parse(string input)
-	{		
-		if (!long.TryParse(input, out var value)) return 0;
-		if (value < 1) return 0;
-		return value;
+	public long Value { get; set; }
+
+	public override string ToString() => $"UKPRN {Value}";
+
+	public static Ukprn? Parse(string input)
+	{
+		if (!long.TryParse(input, out var value)) return null;
+		if (value < 1) return null;
+		return new() { Value = value };
 	}
 }
